Changed file structure for Schema, seeds and select to only be one file, employeeDB.sql

+++ employeeDB.sql
@@ -10,7 +10,7 @@ CREATE TABLE department (
 CREATE TABLE role (
     id INT AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(30) NOT NULL,
-    salary DECIMAL(10,2) NOT NULL,
+    salary INT(10) NOT NULL,
     department_id INT, 
     FOREIGN KEY (department_id) REFERENCES department(id)
 );
